+++ src/components/pagehead.js
@@ -1,8 +1,7 @@
 import React from 'react'
-import { Helmet } from 'react-helmet'
 import { useStaticQuery, graphql } from 'gatsby'
 
-const PageHead = () => {
+const Head = () => {
   const { site } = useStaticQuery(
     graphql`
       query {
@@ -18,19 +17,15 @@ const PageHead = () => {
   )
   return (
     <>
-      <Helmet
-        titleTemplate="%s — Grankontel"
-        defaultTitle={site.siteMetadata?.title}
-      >
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin />
-        <link
-          href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&family=Roboto+Slab&display=swap"
-          rel="stylesheet"
-        />
-      </Helmet>
+      <title>{site.siteMetadata?.title}</title>
+      <link rel="preconnect" href="https://fonts.googleapis.com" />
+      <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin />
+      <link
+        href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&family=Roboto+Slab&display=swap"
+        rel="stylesheet"
+      />
     </>
   )
 }
 
-export default PageHead
+export default Head