+++ src/components/seo.js
@@ -7,7 +7,6 @@
 
  import * as React from "react"
  import PropTypes from "prop-types"
- import { Helmet } from "react-helmet"
  import { useStaticQuery, graphql } from "gatsby"
  
  function Seo({ description, lang, meta, title, image, ogType }) {