+++ src/layouts/page.js
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import PageHead from '../components/pagehead'
+import Head from '../components/pagehead'
 import PageFooter from '../components/pagefooter'
 
 import '../styles/mystyle.scss'
@@ -7,7 +7,7 @@ import '../styles/mystyle.scss'
 const Page = ({ children }) => {
   return (
     <>
-      <PageHead />
+      <Head />
       {children}
       <PageFooter />
     </>