+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.4] - 2023-01-23
+
+- migrated to gatsby 5
+
 ## [1.0.3] - 2021-08-05
 
 - updated project list